% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haida_gwaii}
\alias{haida_gwaii}
\title{Haida Gwaii Data}
\format{An object of class \code{SpatialPolygons} of length 1.}
\usage{
haida_gwaii
}
\description{
Spatial polygon data for Haida Gwaii.
The projection is in Latitude and Longitude for the WGS84 geodetic datum.
}
\details{
The original data are available from \url{http://www12.statcan.gc.ca/census-recensement/2011/geo/bound-limit/files-fichiers/gcd_000b11a_e.zip}.
This does not constitute an endorsement by Statistics Canada of this product.
}
\keyword{datasets}

